moved the fetapi calls to a seperate class
